initial commit, this code works under Ubuntu 14.04
When using this code as separate Code, i.e. removing all 'Overrides' and references to outside classes and making use of a separate Main class, an instance of Control can be created, started and tested. It does then connect to the specified ePuck and executes the commands listed.

To get this code to work certain requirements are necessary:
1) Perform a Desktop-Full installation of ROS 'Indigo' (http://wiki.ros.org/indigo/Installation/Ubuntu)

2) Go through all the steps mentioned in Chapter 3.1 'Initial configuration' of the following document (http://www.gctronic.com/doc/index.php/E-Puck#ROS). Note that instead of step 3 and 4, downloading the epuck_driver from github, I advise to use the epuck_driver folder provided under "additional Files", in the folder 'src' in the archive catkin_ws.zip, as I made some small changes to it.

3) Set the epuck-selector to mode 3.

4) Pair the epuck via bluetooth. The requested PIN is equal to the four-digit number on the epuck.

5) After pairing with the epuck it's MAC-Address can be obtained. Add it in the document ~/catkin_ws/src/epuck_driver/multi_epuck.launch where it belongs.

6) Execute the code.

The class was previously designed to have an getInstance(), probably to be able to retrieve always the same Instance from different places without having to pass the one-time-created Instance around. As my desgin-patterns are a bit rusty I did not implement this for now, but I guess changing some of the fields would easily resolve this issue. Also I think that there should not be that many different places that need an Instace of this class, so maybe we don't need the getInstance()?