rectSearch fixes
---
note: if you want to work with the Pi tomorrow
Getting the HSV values in diff(Mat m1) method is absolutely CRUCIAL.

You can use an online color picker from an image and check the HSV values of the (red) e-puck, of the most different pixels. Then you can create a region that captures all these HSV values (see the diff method, here we use 2 regions but this is not extremely necessary).

Note that OpenCV uses 180, 255, 255 as max values for HSV where most online pickers have 360, 100, 100. So you need to scale the values appropiately!